Moved BmpError and all functions associated with decoding into its own module
